css-align-3/content-distribution/place-content-shorthand-003.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-content shorthand - initial value</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#propdef-place-content" />
<meta name="assert" content="Check that place-content's 'initial' value expands to 'align-content' and 'justify-content'." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-content: start;
        justify-content: end;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    var div = document.getElementById("test");
    div.setAttribute("style", "place-content: initial");
    var style = getComputedStyle(div);

    test(function() {
        assert_equals(style.getPropertyValue("place-content"),
            "normal normal", "place-content resolved value");
    }, "Check place-content: initial - resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("align-content"),
            "normal", "place-content specified value for align-content");
    }, "Check place-content: initial - align-content resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-content"),
            "normal", "place-content specified value for justify-content");
    }, "Check place-content: initial - justify-content resolved value");
</script>
